\name{mPowerStatistics-package}
\alias{mPowerStatistics-package}
\alias{mPowerStatistics}
\docType{package}
\title{
\packageTitle{mPowerStatistics}
}
\description{
\packageDescription{mPowerStatistics}
}
\author{
\packageAuthor{mPowerStatistics}

Maintainer: \packageMaintainer{mPowerStatistics}
}


